requirements.txt
httpx==0.27.*
websockets==12.*
pydantic==2.*
python-dotenv==1.*
rich==13.*
tenacity==9.*
pandas==2.*
uvloop==0.19.*; sys_platform == 'linux'
x10-python-trading-starknet
pytest==8.*
